requirements.txt
flask
pymongo
altair
pandas
requests
pillow
ray